pyproject.toml
[project]
name = "etf-rotation-backtrader"
version = "1.0.0"
description = "ETF Rotation Strategy based on backtrader framework"
authors = [
    {name = "Quant Developer", email = "dev@example.com"}
]
requires-python = ">=3.9"
dependencies = [
    "backtrader>=1.9.78",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "tushare>=1.2.62",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/etf_rotation"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
